+++ tools/cosmovisor/cmd/cosmovisor/main.go
@@ -3,16 +3,22 @@ package main
 import (
 	"context"
 	"os"
+	"time"
 
 	"cosmossdk.io/log"
+
 	cverrors "cosmossdk.io/tools/cosmovisor/errors"
 )
 
 func main() {
-	logger := log.NewLogger(os.Stdout).With(log.ModuleKey, "cosmovisor")
-	ctx := context.WithValue(context.Background(), log.ContextKey, logger)
+	// error logger used only during configuration phase
+	logger := log.NewLogger(
+		os.Stderr,
+		log.ColorOption(false),
+		log.TimeFormatOption(time.Kitchen),
+	).With(log.ModuleKey, "cosmovisor")
 
-	if err := NewRootCmd().ExecuteContext(ctx); err != nil {
+	if err := NewRootCmd().ExecuteContext(context.Background()); err != nil {
 		cverrors.LogErrors(logger, "", err)
 		os.Exit(1)
 	}